# feat: Build cached anchor/alias node list for faster alias resolving