--- a/src/doc/Document.ts
+++ b/src/doc/Document.ts
@@ -433,8 +433,7 @@ export class Document<
       keep: !json,
       mapAsMap: mapAsMap === true,
       mapKeyWarned: false,
-      maxAliasCount: typeof maxAliasCount === 'number' ? maxAliasCount : 100,
-      anchorAndAliasNodes: []
+      maxAliasCount: typeof maxAliasCount === 'number' ? maxAliasCount : 100
     }
     const res = toJS(this.contents, jsonArg ?? '', ctx)
     if (typeof onAnchor === 'function')