--- a/src/nodes/toJS.ts
+++ b/src/nodes/toJS.ts
@@ -10,7 +10,7 @@ export interface AnchorData {
 
 export interface ToJSContext {
   anchors: Map<Node, AnchorData>
-  /** Cached anchor and allias nodes in the order they occurred in the document */
+  /** Cached anchor and alias nodes in the order they occur in the document */
   aliasResolveCache?: Node[]
   doc: Document<Node, boolean>
   keep: boolean