@PolyMeilex 2025-03-29
> Would you be able to share some YAML file with which the performance benefits of the change become measurable?

Had to anonymize it, so it looks kinda goofy, but here is one of the worst offenders in my data set: https://pastebin.com/nPB4hLyT