@PolyMeilex 2025-03-29
>  I'll need to find a bit of time to test how much this improves the performance on it.

Not a proper scientific benchmark, but with the example file this is the difference on my setup:

`before`
```
parse: 141.241ms
toJs: 1.684s
```
`after`
```
parse: 145.326ms
toJs: 26.266ms
```