@PolyMeilex 2025-03-30
We could also build a map where a key would be the anchor's name and the value would be the node with all its child aliases already resolved. Then if a duplicated anchor name is found it would simply replace the previous one, but I'm assuming that there is some reason why that's not how it's done that I'm not aware of.