@PolyMeilex 2025-03-30
> That's effectively `ctx.anchors`, but it's not as simple as one might like. The YAML spec allows for fun structures like this:
> 
> ```yaml
> &foo
> key:
> - *foo
> - &foo 42
> - *foo
> ```
> 
> Here, the resolution of the top-level map depends on resolving all of its contents, and that includes resolving the first `*foo` as a circular reference to the map itself, as well as resolving the second `*foo` as the scalar `42`, thanks to the anchor re-use.
> 
> Getting that to work right was a bit tricky.

Oh, yeah that's fun, would rather pay the price of wasteful iteration then try to implement this 😅

Thank you for being open to this! I will try to find more low hanging fruit so hopefully we can keep generators, while still making the package  usable for my usecase 🚀